+++ .eslintrc.js
@@ -85,6 +85,16 @@ module.exports = {
         message: "Please use src/lib/datetime instead."
       }
     ],
+    "import/no-extraneous-dependencies": [
+      "error",
+      {
+        devDependencies: [
+          "__test__/**/*.ts",
+          "**/*.spec.ts",
+          "webpack.config.js"
+        ]
+      }
+    ],
 
     // Rules to get linting to pass
     camelcase: ["off", { properties: "never" }],